Clean smells

+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/agents/shared/algorithms/distanceCalculator/ManhattanDistance.java
@@ -5,8 +5,8 @@ import com.nscharrenberg.um.multiagentsurveillance.headless.models.Map.Tile;
 public class ManhattanDistance {
 
     public static double compute(Tile tileX, Tile tileY){
-        int x = Math.abs(tileX.getX() - tileY.getX());
-        int y = Math.abs(tileX.getY() - tileY.getY());
+        double x = Math.abs(tileX.getX() - tileY.getX());
+        double y = Math.abs(tileX.getY() - tileY.getY());
         return x + y;
     }
 }
